ipipe: introduce .irq_hold, .irq_release in chip descriptors
We have to differentiate PIC masking operations aimed at deferring
interrupts safely until the root stage can handle them, from other
manipulations of the PIC masking state. Typically, we don't want the
LOCK bit to be set as a result of the former, whilst we do want to
lock/unlock the interrupt in the pipeline log in the latter case.

To this end, we introduce pipeline-specific handlers in the standard
PIC chip descriptor, to hold an IRQ source for deferral (.irq_hold),
and conversely release it (.irq_release).