ipipe/arm: fix arm926 detection
When the ARM SOC used is based on an ARM926-EJS core, disable_hlt should be
called at ipipe init. We used to do this if CONFIG_CPU_ARM926 is defined, but
in order to handle kernel compiled for several cores, we should also detect
at run-time whether the running processor is based on this core. Use a test
on cpu_proc_init to do that.