ipipe/arm: fix SMP cpu_logical_map initialization
The I-pipe patch uses the cpu_logical_map in the "wrong" direction: in
order to obtain the logical cpu id from the physical one. So, in case
where NR_CPUS is smaller than the cpu index of the boot cpu, it will
try to dereference cpu_logical_map(boot_cpu) which will result in an
out of bounds access to the array. Fix that by increasing the array size.