ipipe/x86: implement dynamic tsc selection
When running on a CPU with a TSC, the "rdtsc" instruction is used,
otherwise it is replaced, using the "alternatives" mechanism, with
a call to the the Linux clocksource with the lowest latency.

The default clocksource, which should get picked up when really nothing
else is available, is the PIT. Since the PIT channel 0, used by Linux
PIT clocksource implementation, may also be used as Xenomai hardware
timer, provide an implementation using the PIT channel 2, requiring
to disable the PC speaker, as its driver(s) also normally use
the PIT channel 2.