Fix IRQs-off-tracer for x86_64
Earlier kernels abused the frame pointer to store the
old stack pointer in SAVE_ARGS_IRQ. This behaviour was
changed in upstream commit a2bbe75089d, which breaks
the IRQs-off tracing code for x86_64.

Fix this by using the current stack pointer as basis
to determine the arguments for ipipe_trace_begin/end
and the handler function.

Signed-off-by: Gilles Chanteperdrix <gilles.chanteperdrix@xenomai.org>
Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>
Tested-by: Gernot Hillier <gernot.hillier@siemens.com>