ipipe: Clear IRQ_MASKED from irq descriptor on setup
This bit is set my the kernel on __free_irq. That's unproblematic in the
absence of ipipe, but when it is enabled, handle_fasteoi_irq will fail
to reenable the line on exit due to this spurious state. Only required
on 2.6 kernels, later versions properly clear the MASKED state
themselves.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>