ipipe: Sync head domain IRQ log after domain migration
Typically, ipipe_migration_hook will raise a VIRQ for the head domain in
order to switch over the RT part of the freshly migrated task. As we
call the hook with the head domain stalled, this IRQ cannot be
processed. Ensure that it is not needlessly delayed by flushing IRQs for
the head domain after returning from the hook.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>