x86: Avoid tracing of functions used for early microcode loading
On x86-32, early microcode is loaded before paging is enabled. If we hit
mcount at this point and CONFIG_DYNAMIC_FTRACE is off, we will crash as
the mcount trampoline tries to access global variables at the wrong
addresses.

Avoid this by disabling the instrumentation of the related microcode
files as well as string_32 library which is used as well.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>