x86/ipipe: get FPU backup area allocated for all threads
In order to be able to handle FPU faults in the Xenomai domain without
switching to the root domain, an FPU context must be systematically
allocated for all Linux tasks. Currently, the FPU context is systematically
allocated in arch_dup_task_struct(), but gets freed upon exec by
flush_thread(). Stop freeing it in flush_thread(), there does not seem
to be any risk of leaking information, as the "used_math" flag is cleared
by drop_init_fpu() so that upon next fault, the context is reinitialized.
Also stop creating an FPU context in arch_dup_task_struct() if the source
taks does not have one, as there are no longer any task without an FPU context.