ipipe: introduce the ipipe_smp_p predicate
This predicate allow testing at run-time if we are running on a
multi-processor machine, in order to "comment-out" the useless code when
running an SMP kernel on an UP machine. The default implementation is to
define ipipe_smp_p to be 1 or 0 depending on CONFIG_SMP, but architecture
code can override ipipe_smp_p definition in the CONFIG_SMP case to use
jump labels for instance.