arm/ipipe: fix ipipe_syscall_root with mayday and EABI
In order to pass system call numbers to non-root domains in an uniform way,
ipipe_syscall_root stores the system call number in the r7 register, and
restores the r7 register value after the system call has been handled by
non-root domains. This interferes with Xenomai mayday handling with EABI,
as in the case of a mayday with EABI, in order to restore the context
which was interrupted by the mayday syscall, the r7 register is restored
to the value it had for this interrupted context, but this get overriden
by ipipe_syscall_root handling of the r7 register. What is more, when
running an EABI kernel, it is not necessary to store the syscall number
in r7, and restore it after the syscall, as it is already there.

So, this commit modifies ipipe_syscall_root to only save/restore the r7
register when using OABI. This works because we know that Xenomai mayday
mechanism will not use r7 with OABI kernels. At this chance, optimize
ipipe_syscall_root a little:
- the test for OABI is made at compilation time if possible;
- the irqs disabled warning is only compiled with IPIPE_DEBUG_INTERNAL;
- the value of "current" is saved to a local variable to allow using
it twice without doing twice the SP register masking;
- hard_local_irq_disable/hard_local_irq_enable are replaced with
hard_local_irq_save/hard_local_irq_restore on pre v5 ARM, as this results
in shorter code.

Conflicts:
	arch/arm/kernel/ipipe.c