ipipe: introduce __ipipe_share_current()
This ancillary service may be invoked by client domains for announcing
their intent to share the current task context with the regular
kernel, e.g. like scheduling it.

This is the place to provide for inits that may be required for
enabling such sharing, with a dedicated hook to implement
arch-specific details (__ipipe_arch_share_current()).