ipipe: handle domain downgrading in slow path of IRQ syncer
This fix addresses the case of domain migration which may take place
on behalf of a high priority IRQ handler, from the slow path of the
interrupt dispatcher.

We may have migrated to a different CPU (1) upon return from the
handler, or downgraded from the head domain to the root one (2), the
opposite way is NOT allowed though.

(1) reload the current per-cpu context pointer, so that we further
pull pending interrupts from the proper per-cpu log.

(2) check the stall bit to know whether we may dispatch any interrupt
pending for the root domain, and respin the entire dispatch loop if
so. Otherwise, immediately return to the caller, _without_ affecting
the stall state for the root domain, since we do not own it at this
stage.  This case is basically reflecting what may happen in
dispatch_irq_head() for the fast path.