x86/ipipe: Replicate hw irq state into root domain after exceptions
When we take an exception over the head domain and the handler triggered
a migration to the root domain, we have to ensure that the hardware
interrupt state we found on exception entry is always propagated to the
root domain. We currently only do this when entering over the root
domain.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>