ipipe: protect head domain while updating PTE range protection
In the non-numa case, change_pte_range() will mark each updated PTE in
a range as absent, while the protbits are being fixed up, to make sure
it won't race with the hardware.

We don't want the head domain to preempt the regular kernel until the
updated protbits are written back, so we have to disable hw IRQs
around this section.