x86/ipipe: Restore regs->flags on return from __ipipe_handle_exception
We mangle regs->flags to present them consistently to the Linux handler.
But we have to ensure that this gets fixed up again before returning
from the exception - at least for __ipipe_handle_exception.

__ipipe_divert_exception has no fix-up hook, but we do not run into
troubles as long as we do not take exceptions (that are non-fatal) while
root is stalled and not the head domain. Encode this assumption in form
of a WARN_ON_ONCE.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>