ipipe: Avoid rescheduling from __ftrace_ops_list_func in illegal contexts
The ftrace callback dispatcher may be called with hard irqs disabled or
even over the head domain. We cannot allow any rescheduling in that
case, but also do not have: only non-urgent events are expected to be
kicked-off from ftrace callbacks, and those will at latest be handled on
the next Linux timer tick.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>