arm/ipipe: build CPU reverse map based on DT info
The CPU reverse map is built from the logical map, which in turn
depends on the device tree data, which may select the boot CPU among
any of the available processors. This fix also accounts for the fact
that a cluster ID is part of the MPIDR, which may break the assumption
that entries from the logical map are zero-based, when the cluster ID
is non-zero.

This change does not account for CPUs which belongs to distinct
clusters yet (i.e. we don't index via the MPIDR hash).