powerpc/ipipe: book3s: don't trash stall bit on return from exception
When the pipeline is enabled, the stall bit denotes the current
interrupt state. On the other hand, pt_regs->SOFTE only reflects the
stall bit state on entry of an exception context.

Therefore we should not copy SOFTE back to the stall bit when
returning from exception; the fix drops that part. However we must
mirror the stall bit into SOFTE when entering an exception context
(which we already do properly).