x86/ipipe: Resync exception handling with ipipe-3.14
This is a forward port of 07b11fb8125f and b86e8cbe2288..277a6e00a5b3
from ipipe-3.14. See those commits for the reasons.

The only deviation here is the initialization of flags in case of
unhandled faults over the head domain. We now pick up the flags of the
faulting context instead of use 0 (interrupts off). This seems more
appropriate.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>