arm/ipipe: fix inaccurate RTC on 1 GHz bump
i.MX CPUs usually start with a lower CPU frequency and change to a
higher frequency operating point at kernel start. It has been observed
that when bumping e.g from 792 MHz to 996 MHz during boot, the Linux
clock is counting faster than a RTC. Force the timekeeper to
recalculate the clocksource after the bump. We achieve this by
lowering the rating of the ipipe_tsc and lifting it up again.