ipipe: printk: enable dev_printk() from the head stage
Calling dev_{info,warn,err}() is currently invalid, as vprintk_emit()
eventually operates on regular kernel resources.

If called from the head stage, or with hard irqs disabled in presence
of a head stage, route all dev_printk() requests to our deferred
logging mechanism already used for printk() instead. This means that
we won't syslog the output in such a case, but at least we will have a
feedback on the console device, without doing anything silly in the
meantime.

[Jan: back-ported to 4.4]
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>