ipipe: get back to simplistic idle state control
Wait for the interface rework ongoing in I-pipe/4.14 before asking the
co-kernel to have its say again, based on more information about the
cpuidle device.

For now, idling is denied only if IRQs are pending for the root stage.

At this chance, a redundant execution of the idling check was
fixed.