ARM: ipipe: fix section mismatches for KUSER_TSC helpers
The change in #5a7e6eeede58 does not fix the root issue, but causes
__ipipe_tsc_register() to refer to .init data which is a section
violation too. Unfortunately, __ipipe_tsc_register() may be called
long after boot when loading a clocksource dynamically (which is what
the above commit was trying to address), but the position of TSC
helpers in the KUSER page depends on other -unrelated- helpers so we
have to define them in the same section.

To escape this stalemate, move the whole KUSER setup code and
variables outside of the .init section (TSC and other helpers), which
should only amount to a few hundreds of bytes.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>