x86: ipipe: Harden racy path in __fpu__restore_sig
This is needed because I-pipe can interrupt at a point where
fpstate_active is already set but the (internally hardened)
fpu__restore() was not run yet. As the context switch uses
fpstate_active to decide whether to activate the FPU for the target
thread, we may prematurely activate it. That causes at least a warning
by __fpregs_activate() about duplicate activation, but maybe worse
things can happen.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>