ipipe: x86: Do not call BUG() in case of retriggered APIC vectors
The CPU hotplugging code might re-trigger a vector. In that case the
vector is in the VECTOR_RETRIGGERED state. With this patch applied ipipe
does no longer call BUG() for such vectors.

This is a backport from the 4.19 and 5.4 ipipe branches handling the
VECTOR_SHUTDOWN state. The VECTOR_SHUTDOWN case itself does not exist in
4.4, but the VECTOR_RETRIGGERED case does.

__ipipe_handle_irq() is now synchronized with do_IRQ().

Signed-off-by: Florian Bezdeka <florian.bezdeka@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>